pyproject.toml
[project]
name = "erleah-backend"
version = "0.1.0"
description = "AI-powered conference assistant backend with LangGraph"
authors = [{name = "Richard", email = "richard@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv>=1.0.0",
    
    # Agent framework (LangGraph)
    "langgraph>=0.0.25",
    "langchain>=0.1.6",
    "langchain-anthropic>=0.1.4",
    
    # Vector database
    "qdrant-client>=1.7.3",
    
    # Relational database
    "psycopg2-binary>=2.9.9",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    
    # Cache
    "redis>=5.0.1",
    
    # Utilities
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    "sse-starlette>=1.8.2",

    # Structured logging
    "structlog>=24.1.0",

    # Config
    "pyyaml>=6.0",

    # Monitoring
    "prometheus-client>=0.20.0",

    # xAI / Grok (acknowledgments)
    "openai>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "black>=23.12.1",
    "ruff>=0.1.11",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]
